+++ demo/zfsrc/ZFFramework_test/ZFAlgorithm_test/ZFAlgorithm_ZFCompress_test.cpp
@@ -25,7 +25,7 @@ protected:
         this->testCaseOutput("compress buffer");
         {
             ZFIOBufferedCallbackUsingBuffer io;
-            ZFCompress(io, ZFInputForBuffer("uncompressed text"));
+            ZFCompress(io, ZFInputForBufferUnsafe("uncompressed text"));
             ZFDecompress(ZFOutputDefault(), io);
         }